% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_convert.Internal.R
\name{back_convert.Internal}
\alias{back_convert.Internal}
\title{Internal function, not for use. Used to convert from numeric output to AGCT.}
\usage{
back_convert.Internal(sim_samples, allele_freq_data)
}
\arguments{
\item{list}{outputted from the \link{sim_single_sample} function.}

\item{tibble}{ouputted from the \link{read_in_af_data} function.}
}
\value{
A list of \code{n_samples} elements, each corresponding to a sample. Each element is a matrix with a row for each allele, and a column for each SNP. The elements of the matrix are characters for nucleotides (A,C,G,T).
}
\description{
The output of the \link{sim_single_sample} function is integers, which index the variants of each of the SNPs. This function uses the allele frequency data-frame to convert these integers back to the variants (A,G,C,T).
}
