% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allele_frequency_input.R
\name{check_allele_frequency_input}
\alias{check_allele_frequency_input}
\title{Internal file which checks the input file for errors and misspecifications.}
\usage{
check_allele_frequency_input(allele_freq_data, ancestor_pop_label = NULL)
}
\arguments{
\item{allele_freq_data}{Direct read-in from the input data .csv file.}

\item{ancestor_pop_label}{A vector of integers giving the highest level of the family tree, which consists only of ancestors from pure genetic populations. Integers index the population, with order given by the order in which the population dominant allele frequencies are provided in the input allele frequencies file.}
}
\description{
This function checks the input file of allele frequencies and makes sure the required columns are both present in the correct format.
}
