% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_finder.R
\name{genotype_finder}
\alias{genotype_finder}
\title{Internal function which finds possible genotypes from available SNP variants.}
\usage{
genotype_finder(allele_freq_data)
}
\arguments{
\item{allele_freq_data}{Processed allele frequencies data-set read in using \link{read_in_af_data}.}
}
\value{
A list of possible genotypes based on allele variants.
}
\description{
Internal function which finds possible genotypes from available SNP variants.
}
