% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_af_data.R
\name{read_in_af_data}
\alias{read_in_af_data}
\title{Read in the population allele frequencies from prespecified CSV file format.}
\usage{
read_in_af_data(file, ancestor_pop_label = NULL)
}
\arguments{
\item{file}{Path to CSV file containing the population allele frequencies.}

\item{ancestor_pop_label}{A vector of integers giving the highest level of the family tree, which consists only of ancestors from pure genetic populations. Integers index the population, with order given by the order in which the population dominant allele frequencies are provided in the input allele frequencies file. In this function this parameter is only used to check that it corresponds correctly to the input file. Leaving it at its default value (`NULL`) will bypass these checks and load the input file.}
}
\value{
A \link[tibble]{tibble} data-frame with a column for the SNP ID called \code{snp_id}, a column for the variants called \code{variants} which contain character vectors for each SNP, and a column for each population of allele frequencies as sum-to-one constrained vectors.
}
\description{
This function is used internally to read in allele frequency information and complete basic cleaning. To ensure that the correct probabilities are used in simulation, care should be taken to ensure the input file complies with the required format (see Details below), and that the order of the populations given in the input file is kept consistent with the input \code{ancestor_pop_label} in the simulation functions.
}
\details{
The CSV file must comply with the following format rules:
\itemize{
\item{Each row corresponds to a SNP.}
\item{The file's first 3 columns must be the SNP ID, The chromosome and position, and the SNP variants.}
\item{ The variants must be specified by a list of nucleotides separated by backslashes '/', where the most frequent allele is first. E.g. 'A/G' specifies a SNP with the most common variant being 'A' and the least common being 'G'.}
\item{ In the following columns the derived allele frequencies in each pure population. The column names must have "(D)" at the end of their name to be recognised.}
\item{ Only ancestor population allele frequencies can have names ending with (D).}
}
}
\examples{
#Download file to temporary directory:
temp_dir <- tempdir()
download.file(url = "https://raw.githubusercontent.com/danwkenn/SimAdmixtR/master/inst/example-files/input-allele-frequencies-three-snps.csv",destfile = paste0(temp_dir,"/input-af-example.csv"))
allele_frequencies.tbl <- read_in_af_data(file = paste0(temp_dir,"/input-af-example.csv"))
allele_frequencies.tbl
}
