% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_App.R
\name{run_App}
\alias{run_App}
\title{Run a basic User Interface to easily run simulation experiments and "download" the simulated data.}
\usage{
run_App()
}
\description{
Run a basic User Interface to easily run simulation experiments and "download" the simulated data.
}
\examples{
\dontrun{run_App()}
}
