% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single_sample.R
\name{sim_single_sample}
\alias{sim_single_sample}
\title{Function which simulates inheritance for a set of genes for a single simulated person.}
\usage{
sim_single_sample(ancestor_pop_label, allele_freq_list)
}
\arguments{
\item{ancestor_pop_label}{A vector of integers giving the highest level of the family tree, which consists only of ancestors from pure genetic populations. Integers index the population, with order given by the order in which the population dominant allele frequencies are provided in the input allele frequencies file.}

\item{allele_freq_list}{A list of allele frequencies for the ancestor populations. The upper level must be SNP, and the lower level corresponds to population frequencies in order of the population indexes in \code{ancestor_pop_label}.}
}
\description{
Function which simulates inheritance for a set of genes for a single simulated person.
}
