% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_admixture.R
\name{simulate_admixture}
\alias{simulate_admixture}
\title{Simulate genetic inheritance of SNPs based on population allele frequencies and family tree structure.}
\usage{
simulate_admixture(n_samples, ancestor_pop_label, file)
}
\arguments{
\item{n_samples}{The number of samples to simulate.}

\item{ancestor_pop_label}{A vector of integers giving the highest level of the family tree, which consists only of ancestors from pure genetic populations. Integers index the population, with order given by the order in which the population dominant allele frequencies are provided in the input allele frequencies file.}

\item{file}{File with input allele frequencies of ancestor populations. This file must be in a specific format as exemplified in `input-af-example.csv`. For a complete description of the format see \code{\link{read_in_af_data}}.}
}
\value{
A list of \code{n_samples} elements, each corresponding to a sample. Each element is a matrix with a row for each allele, and a column for each SNP. The elements of the matrix are characters for nucleotides (A,C,G,T).
}
\description{
Simulate genetic inheritance of SNPs based on population allele frequencies and family tree structure.
}
\examples{
#Download file to temporary directory:
temp_dir <- tempdir()
download.file(url = "https://raw.githubusercontent.com/danwkenn/SimAdmixtR/master/inst/example-files/input-allele-frequencies-three-snps.csv",destfile = paste0(temp_dir,"/input-af-example.csv"))
sim_data <- simulate_admixture(
n_samples = 10,
ancestor_pop_label = c(1,2,2,2),
file = paste0(temp_dir,"/input-af-example.csv")
)
}
