% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Simple function to pair up the elements of two or more equal-length lists. Function works similar to zip in python.}
\usage{
stitch(list1, ...)
}
\arguments{
\item{list1}{A list.}

\item{...}{More lists of the same length.}
}
\value{
A list of the same length as the original lists. Each element of the list is itself a list, containing two elements: the corresponding element of the first list, and the corresponding element of the second list.
The function should will also work with vectors of equal length.
}
\description{
Simple function to pair up the elements of two or more equal-length lists. Function works similar to zip in python.
}
\examples{
listA = list(1,2,3,4)
listB = list("A","B","C","D")
stitch(listA,listB)
}
