% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_vector.R
\name{string_to_vector}
\alias{string_to_vector}
\title{Simple internal function to create a vector of characters from a string.}
\usage{
string_to_vector(x)
}
\arguments{
\item{x}{A string.}
}
\value{
a vector of characters.
}
\description{
Simple internal function to create a vector of characters from a string.
}
