% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_admixture.R
\name{summarise_admixture}
\alias{summarise_admixture}
\title{Function finds the genotype frequencies for a simulated data-list, and the most likely genotype for each SNP.}
\usage{
summarise_admixture(data_list, pop_allele_file)
}
\arguments{
\item{data_list}{Simulated data output from \link{simulate_admixture}.}

\item{pop_allele_file}{Path to CSV file containing the population allele frequencies.}
}
\value{
A list containing:
\itemize{
\item{\code{frequencies}: A list of vectors, with one element for each SNP. The vectors have a single element for each possible genotype, and the value is the observed frequency of that genotype in the simulated data.}
\item{\code{most_common}: a matrix of the most common genotype found in the sample.}
}
}
\description{
Function finds the genotype frequencies for a simulated data-list, and the most likely genotype for each SNP.
}
