% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_structure.R
\name{write_to_structure}
\alias{write_to_structure}
\title{Take simulated genotypes and write to a STRUCTURE input file format.}
\usage{
write_to_structure(
  sim_data,
  structure_snp_order = NULL,
  example_structure_file = NULL,
  sample_names = NULL,
  output = "test",
  type = c("txt", "csv"),
  keep_extra_snps = TRUE
)
}
\arguments{
\item{sim_data}{Output from simulate_admixture}

\item{structure_snp_order}{(Optional) The desired order of SNPs for columns of output file.}

\item{example_structure_file}{(Optional) A sample STRUCTURE input file to extract the desired order of SNPs from. If both \code{example_structure_file} and \code{structure_snp_order} are specified, the order given by this will take precedence.}

\item{sample_names}{(Optional) vector of names for the simulated samples.}

\item{output}{Name for the output file.}

\item{type}{Options for txt or csv file outputs.}

\item{keep_extra_snps}{If there are SNPs present in the simulated data, but not in the SNP order provided by \code{example_structure_file} or \code{structure_snp_order}, should the extra SNPs be kept and appended to the table (on the right)?}
}
\value{
\code{NULL}
}
\description{
Take simulated genotypes and write to a STRUCTURE input file format.
}
\examples{
#Download files to temporary directory:
temp_dir <- tempdir()
download.file(url = "https://raw.githubusercontent.com/danwkenn/SimAdmixtR/master/inst/example-files/input-allele-frequencies-three-snps.csv",destfile = paste0(temp_dir,"/input-af-example.csv"))
download.file(url = "https://raw.githubusercontent.com/danwkenn/SimAdmixtR/master/inst/example-files/simple-three-allele-example.csv",destfile = paste0(temp_dir,"/example-structure-file.csv"))

sim_data <- simulate_admixture(
n_samples = 10,
ancestor_pop_label = c(1,2,2,2),
file = paste0(temp_dir,"/input-af-example.csv")
)

write_to_structure(
sim_data = sim_data,
example_structure_file = paste0(temp_dir,"/example-structure-file.csv"),
output = "example-output",
type = c("txt","csv"))
}
