% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_finder.R
\name{genotype_finder}
\alias{genotype_finder}
\title{Internal function which finds possible genotypes from available SNP variants.}
\usage{
genotype_finder(allele_freq_data)
}
\description{
Internal function which finds possible genotypes from available SNP variants.
}
